Clean up some of the existing Balance Event code

+++ backend/Tests/DatabaseFixture.cs
@@ -44,7 +44,7 @@ public class DatabaseFixture : IDisposable
         {
             return;
         }
-        if (disposing)
+        if (disposing && ShouldUseDatabase)
         {
             Directory.Delete(DatabaseDirectory, true);
         }
